chore(inline-loading): storybook controls
