+++ packages/react/src/components/InlineLoading/InlineLoading.mdx
@@ -1,4 +1,5 @@
-import { Props } from '@storybook/addon-docs';
+import { Props, Preview, Story } from '@storybook/addon-docs';
+import InlineLoading from '../InlineLoading';
 
 # InlineLoading
 
@@ -11,6 +12,10 @@ import { Props } from '@storybook/addon-docs';
 ## Table of Contents
 
 ## Overview
+
+<Preview>
+  <Story id="components-inlineloading--default" />
+</Preview>
 
 ## Component API
